Preloading images on landing
Using a new controller to preload project card images to ensure smooth transition instead of having the transition hop around as it fetches images.  This 'adds load time' to the intro page, but it is imperceivable to the user since the core intro actions are unaffected by the asychronous nature of the image fetching going on in the background.